requirements.txt
# Core Scientific Stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Chemistry and Cheminformatics
rdkit
chembl_webresource_client>=0.10.8
pyarrow>=12.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data and File Handling
openpyxl>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and Reporting
loguru>=0.7.0
tabulate>=0.9.0

# Model Persistence
joblib>=1.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Interpretability
shap>=0.44.0
